Working version
I got stuck on the behaviour of the rubberband and reached out on the GIS Stack Exchange for some assistance. A member there made some edits that provided the ability to roate the map and rubber band accordingly. The member took a different approach than what I did so none of this code has anything in it that I did originally and I cannot take any credit. I am only posting it to share for those who were also looking for this feature. Using this approach I am working on adding it to the 2.X version of the plugin.